CodeMirror fixes after adding GraphiQL; split CodeMirror editor component into its own chunk
