+++ app/javascript/packs/application.js
@@ -16,27 +16,11 @@ import '../displayBrowserWarning'; // runs on import
 import Rails from 'rails-ujs';
 import 'bootstrap.native/dist/bootstrap-native-v4';
 
-import 'codemirror';
-import 'codemirror/mode/markdown/markdown';
-import 'codemirror/mode/htmlmixed/htmlmixed';
-import 'codemirror/mode/javascript/javascript';
-import 'codemirror/addon/mode/multiplex';
-import 'codemirror/lib/codemirror.css';
-import 'codemirror/addon/edit/matchbrackets';
-import 'codemirror/addon/edit/matchtags';
-import 'codemirror/addon/fold/foldcode';
-import 'codemirror/addon/fold/brace-fold';
-import 'codemirror/addon/fold/markdown-fold';
-import 'codemirror/addon/fold/xml-fold';
-import 'codemirror/addon/fold/foldgutter';
-import 'codemirror/addon/fold/foldgutter.css';
-
 import WebpackerReact from 'webpacker-react';
 
 import '../styles/application.scss';
 import 'react-table/react-table.css';
 import '../inflections';
-import '../Codemirror/LiquidMultiplexModes';
 
 import components from './components';
